Title: 10개 입력받기, Time: 137ms, Memory: 10MB, Status: Wrong Answer - Codetree

+++ 251205/10개 입력받기/receive-10-inputs.java
@@ -18,7 +18,7 @@ public class Main {
             if(arr[i] == 0)
                 break;
         }
-        float avg = (float)(sum /(cnt-1));
+        double avg = (double)(sum /(cnt-1));
         System.out.print(sum + " ");
         System.out.printf("%.1f",avg);
     }
